Fix track lookup in the platform specific code for Android and iOS
All operations on track objects in the flutter API just send the
track id to the platform specific code without the additional
information which peer connection the track belongs to.
This leads to methods like captureFrame() being called on the
wrong track if there are tracks with the same id in more than one
peer connection.

Fixes #469